requirements.txt
Django>=2.0
psycopg2
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
paramiko==2.4.2
uwsgi>=2.0
Pillow==5.3.0
minio==4.0.8